Added options for height - under 5', 5'-5'6", 5'6"-6' over 6' (words written out for clarity) instead of the text entry
Added special requirements note area with text suggesting things like step through, rack - small bike, add kids, helmets, locks and lights, coaster brakes

Add a ratio of ready to appointments - meaning of the bikes ready for clients, there must be 1.5 available. For example if there are 7 bikes, only 5 clients can make appointments (I'm a little light on the math, please help me).

Restored the warning on the page if no bikes are available and prevent people from requesting bikes when we don't have any

Added flag that indicates the coop is closed for the season

Added bike assignment to make it easier to associate a bike with a client